orm maps to azure sql now

+++ BE/src/main/java/com/g3/feedbackApp/Repository/IUserRepository.java
@@ -3,7 +3,7 @@ package com.g3.feedbackApp.Repository;
 import com.g3.feedbackApp.Models.UserModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUserRepository extends JpaRepository<UserModel, Integer> {
+public interface IUserRepository extends JpaRepository<UserModel, Long> {
 
     UserModel getUserModelByStudentNr(Long nr);
     UserModel getUserModelByEmail(String email);
